requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
authlib==1.2.0
SQLAlchemy==2.0.25
jinja2==3.1.2
aiofiles==23.1.0
python-dotenv==1.0.0